pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["marinaRomanchuk <marina.romanchuk23@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "3.2.11"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"
Pillow = "^9.0.1"
drf-extra-fields = "^3.2.1"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
pylama = "^8.3.7"
mypy = "^0.931"
black = "21.12b0"
pre-commit = "^2.17.0"
coverage = "^6.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
